# server/docs/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "Valley MAS 创作者内容平台 API 文档\n支持创作者注册、口令管理、资源上传下载等功能",
    "title": "Valley MAS API",
    "contact": {
      "name": "API Support",
      "email": "support@valley-mas.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    },
    "version": "1.0"
  },
  "host": "localhost:8080",
  "basePath": "/api/v1",
  "paths": {
...
    "/creator/register": {
      "post": {
        "security": [
          {
            "Bearer": []
          }
        ],
        "description": "普通用户注册成为创作者，系统自动生成5位口令（如：y2722）",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "tags": ["创作者"],
        "summary": "注册成为创作者",
        "parameters": [
          {
            "description": "创作者信息",
            "name": "request",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/handler.RegisterCreatorRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "注册成功",
            "schema": {
              "type": "object",
              "additionalProperties": true
            }
          },
          "400": {
            "description": "参数错误或已注册",
            "schema": {
              "type": "object",
              "additionalProperties": true
            }
          },
          "401": {
            "description": "未登录",
            "schema": {
              "type": "object",
              "additionalProperties": true
            }
          },
          "500": {
            "description": "服务器错误",
            "schema": {
              "type": "object",
              "additionalProperties": true
            }
          }
        }
      }
    }
  },
  "definitions": {
    "handler.RegisterCreatorRequest": {
      "type": "object",
      "required": ["name"],
      "properties": {
        "avatar": {
          "type": "string",
          "example": "https://example.com/avatar.jpg"
        },
        "description": {
          "type": "string",
          "maxLength": 255,
          "example": "分享精美头像和壁纸"
        },
        "name": {
          "type": "string",
          "maxLength": 50,
          "minLength": 2,
          "example": "设计师小王"
        },
        "spaceBanner": {
          "type": "string",
          "example": "https://example.com/banner.jpg"
        },
        "spaceDescription": {
          "type": "string",
          "example": "这里有最新的设计作品"
        },
        "spaceTitle": {
          "type": "string",
          "maxLength": 100,
          "example": "小王的创意空间"
        }
      }
    },
...
  },
  "securityDefinitions": {
    "Bearer": {
      "description": "JWT 认证，格式：Bearer {token}",
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  }
}